rename read dir to query
