+++ cardinal/main.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 
 	"github.com/argus-labs/starter-game-template/cardinal/component"
-	"github.com/argus-labs/starter-game-template/cardinal/read"
 	"github.com/argus-labs/starter-game-template/cardinal/system"
 	"github.com/argus-labs/starter-game-template/cardinal/tx"
 	"github.com/rs/zerolog"
@@ -50,7 +49,7 @@ func main() {
 	// NOTE: You must register your read endpoints here,
 	// otherwise it will not be accessible.
 	err = cardinal.RegisterQueries(world,
-		read.Constant,
+		query.Constant,
 	)
 	if err != nil {
 		log.Fatal().Err(err)
